+++ Exodus/src/utopia/exodus/rest/resource/description/LanguagesNode.scala
@@ -1,11 +1,10 @@
 package utopia.exodus.rest.resource.description
 
-import utopia.citadel.database.access.many.description.DbLanguageDescriptions
 import utopia.citadel.database.access.many.language.DbLanguages
 import utopia.exodus.rest.util.AuthorizedContext
+import utopia.metropolis.model.cached.LanguageIds
 import utopia.metropolis.model.combined.language.DescribedLanguage
 import utopia.metropolis.model.enumeration.ModelStyle.Simple
-import utopia.metropolis.model.stored.description.DescriptionLink
 import utopia.metropolis.model.stored.language.Language
 import utopia.nexus.result.Result
 import utopia.vault.database.Connection
@@ -33,12 +32,6 @@ class LanguagesNode(authorization: (AuthorizedContext, => Result, Connection) =>
 	override protected def authorize(onAuthorized: => Result)(implicit context: AuthorizedContext, connection: Connection) =
 		authorization(context, onAuthorized, connection)
 	
-	override protected def items(implicit connection: Connection) = DbLanguages.all
-	
-	override protected def descriptionsAccess = DbLanguageDescriptions
-	
-	override protected def idOf(item: Language) = item.id
-	
-	override protected def combine(item: Language, descriptions: Set[DescriptionLink]) =
-		DescribedLanguage(item, descriptions)
+	override protected def describedItems(implicit connection: Connection, languageIds: LanguageIds) =
+		DbLanguages.described
 }